Allow fractional number data

+++ src/DataAdd.js
@@ -63,7 +63,7 @@ const DataAdd = () => {
     if (!nextData.date) nextData.date = new Date().toISOString();
     if (!nextData.name) nextData.name = category.name;
     if (category.type === 'number')
-      nextData.value = parseInt(nextData.value, 10);
+      nextData.value = parseFloat(nextData.value, 10);
     if (category.type === 'yes/no') nextData.value = true;
     nextTrack.data.unshift(nextData);
     sortOn(nextTrack.data, 'date', 'desc');
